+++ CHANGELOG.md
@@ -7,8 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.1] - 2025-06-18
-
 ### Added
 - Response examples for all endpoints
 
@@ -16,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pre-request script for rrset delete
 - Publish workflow
   - The collection was causing arg limit errors with `jq`
+  - Wrote a simple Python script to handle the publishing to Postman
 
 ### Removed
 - Subaccount section